# *: return *InternalKV from internal iterators