--- a/compaction_test.go
+++ b/compaction_test.go
@@ -966,8 +966,8 @@ func TestCompaction(t *testing.T) {
 
 	get1 := func(iter internalIterator) (ret string) {
 		b := &bytes.Buffer{}
-		for key, _ := iter.First(); key != nil; key, _ = iter.Next() {
-			b.Write(key.UserKey)
+		for kv := iter.First(); kv != nil; kv = iter.Next() {
+			b.Write(kv.UserKey)
 		}
 		if err := iter.Close(); err != nil {
 			t.Fatalf("iterator Close: %v", err)