This is a very large, fundamental change to the iterator stack, so I'd like to give it plenty of discussion. I put the PR together ahead of time so that we could clearly see what it looks like, but I'm happy to discard it if we decide against it. Also, speaking of large PRs, this one is truly massive. However, it is purely mechanical, and almost all of it trivially so. The levelIter's `smallestBoundary`,`largestBoundary` fields required maybe the least trivial changes.

If we decide to move forward and it would ease review, I'm also happy to split it across multiple PRs using adapter types to bridge interface gaps.

----

Previously, internal iterators' positioning methods returned two return values, a pointer to an internal key and a LazyValue struct. This commit collapses these two return values into a single return value: a pointer to an InternalKV struct. The key and value buffers were already owned by the underlying iterator, so passing the base.LazyValue as a value did not decouple the two. Propagating a single *InternalKV pointer can be cheaper in some instnaces.

More importantly, it allows us to propagate additional information without increasing the amount of copying up and down the iterator stack. I anticipate this being useful in the implementation of cockroachdb/cockroach#111001, allowing us to propagate checksums alongside keys and values. Merging this change before the branch cut will reduce the pain of any backports.

Iterator microbenchmarks show movement in both directions, mostly ± ~4%.

```
goos: linux
goarch: amd64
pkg: github.com/cockroachdb/pebble
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                                                                                                                │   old.txt    │                new.txt                 │
                                                                                                                │    sec/op    │    sec/op     vs base                  │
IteratorSeekGE-24                                                                                                 714.6n ±  2%   747.1n ±  3%   +4.54% (p=0.000 n=10)
IteratorNext-24                                                                                                   26.18n ±  0%   25.53n ±  0%   -2.50% (p=0.000 n=10)
IteratorPrev-24                                                                                                   35.49n ±  0%   36.48n ±  1%   +2.78% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=false-24                        503.8n ±  0%   497.8n ±  0%   -1.18% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=true-24                         388.8n ±  0%   384.8n ±  0%   -1.02% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=false-24                         1.070µ ±  0%   1.082µ ±  0%   +1.17% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=true-24                          465.7n ±  1%   467.9n ±  1%   +0.48% (p=0.009 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=false-24                         582.6n ±  0%   559.3n ±  0%   -4.00% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=true-24                          403.8n ±  0%   399.6n ±  0%   -1.03% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=false-24                          1.093µ ±  0%   1.103µ ±  1%   +0.87% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=true-24                           464.9n ±  1%   471.7n ±  0%   +1.46% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=false-24                        503.7n ±  0%   497.9n ±  0%   -1.15% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=true-24                         389.1n ±  0%   384.4n ±  0%   -1.21% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=false-24                         1.068µ ±  1%   1.085µ ±  1%   +1.54% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=true-24                          463.1n ±  1%   463.9n ±  1%        ~ (p=0.631 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=false-24                         583.0n ±  0%   559.1n ±  0%   -4.08% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=true-24                          403.4n ±  0%   399.6n ±  0%   -0.94% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=false-24                          1.094µ ±  1%   1.103µ ±  0%   +0.82% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=true-24                           463.2n ±  1%   462.9n ±  0%        ~ (p=0.305 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=false-24                        504.9n ±  0%   498.9n ±  0%   -1.19% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=true-24                         389.6n ±  0%   385.6n ±  0%   -1.03% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=false-24                         1.060µ ±  0%   1.065µ ±  0%   +0.42% (p=0.012 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=true-24                          461.8n ±  0%   460.2n ±  0%   -0.36% (p=0.041 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=false-24                         584.1n ±  0%   560.3n ±  0%   -4.08% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=true-24                          404.3n ±  0%   400.7n ±  0%   -0.90% (p=0.000 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=false-24                          1.085µ ±  0%   1.090µ ±  1%   +0.46% (p=0.003 n=10)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=true-24                           463.4n ±  0%   462.0n ±  1%        ~ (p=0.255 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=false-24                           605.7n ±  0%   593.3n ±  0%   -2.03% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=true-24                            489.3n ±  0%   483.3n ±  0%   -1.24% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=false-24                            1.269µ ±  0%   1.243µ ±  1%   -2.01% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=true-24                             755.5n ±  0%   742.8n ±  0%   -1.69% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=false-24                            683.7n ±  0%   663.3n ±  0%   -2.98% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=true-24                             508.2n ±  0%   498.3n ±  0%   -1.94% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=false-24                             1.268µ ±  0%   1.238µ ±  1%   -2.41% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=true-24                              776.4n ±  0%   754.6n ±  1%   -2.81% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=false-24                           641.5n ±  0%   629.6n ±  0%   -1.86% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=true-24                            521.8n ±  0%   514.2n ±  0%   -1.46% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=false-24                            1.302µ ±  0%   1.282µ ±  0%   -1.57% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=true-24                             783.1n ±  0%   767.3n ±  0%   -2.02% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=false-24                            721.1n ±  0%   700.3n ±  0%   -2.88% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=true-24                             540.3n ±  0%   529.8n ±  0%   -1.93% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=false-24                             1.305µ ±  0%   1.272µ ±  0%   -2.53% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=true-24                              803.2n ±  0%   779.1n ±  1%   -3.00% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=false-24                           709.9n ±  0%   694.5n ±  0%   -2.16% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=true-24                            588.5n ±  0%   578.8n ±  0%   -1.65% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=false-24                            1.377µ ±  0%   1.348µ ±  1%   -2.07% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=true-24                             842.3n ±  0%   824.4n ±  1%   -2.13% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=false-24                            791.5n ±  0%   768.2n ±  0%   -2.94% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=true-24                             610.2n ±  0%   595.6n ±  0%   -2.39% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=false-24                             1.377µ ±  0%   1.341µ ±  0%   -2.58% (p=0.000 n=10)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=true-24                              861.5n ±  1%   841.1n ±  1%   -2.37% (p=0.000 n=10)
IteratorSeqSeekGEWithBounds/two-level=false-24                                                                    912.5n ±  1%   893.6n ±  1%   -2.07% (p=0.000 n=10)
IteratorSeqSeekGEWithBounds/two-level=true-24                                                                     911.4n ±  0%   893.6n ±  0%   -1.95% (p=0.000 n=10)
IteratorSeekGENoop/withLimit=false-24                                                                             45.12n ±  1%   50.98n ±  0%  +12.98% (p=0.000 n=10)
IteratorSeekGENoop/withLimit=true-24                                                                              82.73n ±  0%   89.69n ±  0%   +8.41% (p=0.000 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward/seekprefix-24                                            45.07m ±  5%   44.94m ±  3%        ~ (p=0.393 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward/next-24                                                  4.023m ±  1%   3.974m ±  1%   -1.23% (p=0.019 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-24                                                      4.877m ±  1%   4.843m ±  1%        ~ (p=0.247 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward/seekprefix-24                                            44.63m ±  5%   44.79m ±  5%        ~ (p=0.579 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward/next-24                                                  3.584m ±  1%   3.546m ±  1%   -1.04% (p=0.002 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-24                                                      4.188m ±  1%   4.213m ±  2%        ~ (p=0.796 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward/seekprefix-24                                            44.41m ±  3%   45.65m ±  2%   +2.79% (p=0.035 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward/next-24                                                  2.109m ±  1%   2.115m ±  1%        ~ (p=0.579 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-24                                                      2.700m ±  1%   2.744m ±  1%   +1.64% (p=0.000 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward/seekprefix-24                                           43.51m ±  2%   44.56m ±  5%   +2.40% (p=0.029 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward/next-24                                                 323.5µ ±  0%   317.1µ ±  1%   -2.00% (p=0.000 n=10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-24                                                     621.6µ ±  1%   621.2µ ±  1%        ~ (p=0.739 n=10)
IteratorScan/keys=100,r-amp=1,key-types=points-only-24                                                            9.158µ ±  2%   9.227µ ±  1%        ~ (p=0.315 n=10)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-24                                                      9.680µ ± 20%   9.619µ ±  1%        ~ (p=0.481 n=10)
IteratorScan/keys=100,r-amp=3,key-types=points-only-24                                                            16.13µ ±  5%   16.62µ ± 16%   +3.06% (p=0.029 n=10)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-24                                                      16.54µ ± 42%   16.54µ ±  0%        ~ (p=0.869 n=10)
IteratorScan/keys=100,r-amp=7,key-types=points-only-24                                                            23.16µ ±  8%   24.06µ ± 18%   +3.92% (p=0.019 n=10)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-24                                                      23.68µ ±  1%   23.98µ ±  2%   +1.29% (p=0.005 n=10)
IteratorScan/keys=100,r-amp=10,key-types=points-only-24                                                           28.10µ ±  1%   28.39µ ± 24%   +1.05% (p=0.005 n=10)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-24                                                     28.68µ ± 23%   28.78µ ± 23%        ~ (p=0.353 n=10)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-24                                                           72.03µ ±  0%   72.78µ ±  1%   +1.04% (p=0.002 n=10)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-24                                                     75.56µ ±  0%   75.98µ ±  1%   +0.56% (p=0.000 n=10)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-24                                                           119.7µ ±  0%   121.9µ ±  0%   +1.82% (p=0.000 n=10)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-24                                                     123.8µ ±  0%   125.7µ ±  0%   +1.54% (p=0.000 n=10)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-24                                                           153.1µ ±  1%   155.0µ ±  0%   +1.23% (p=0.002 n=10)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-24                                                     156.8µ ±  1%   159.3µ ±  1%   +1.55% (p=0.002 n=10)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-24                                                          164.6µ ±  0%   167.6µ ±  0%   +1.82% (p=0.000 n=10)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-24                                                    169.3µ ±  1%   171.9µ ±  0%   +1.51% (p=0.001 n=10)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-24                                                          706.3µ ±  0%   717.6µ ±  4%   +1.60% (p=0.001 n=10)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-24                                                    739.2µ ±  1%   751.0µ ±  0%   +1.59% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-24                                                          1.155m ±  0%   1.180m ±  0%   +2.15% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-24                                                    1.195m ±  0%   1.219m ±  0%   +2.01% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-24                                                          1.441m ±  0%   1.467m ±  0%   +1.76% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-24                                                    1.481m ±  0%   1.507m ±  0%   +1.80% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-24                                                         1.526m ±  0%   1.569m ±  1%   +2.79% (p=0.000 n=10)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-24                                                   1.573m ±  0%   1.609m ±  1%   +2.34% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=false/key-types=points-only-24              143.5n ±  0%   140.9n ±  0%   -1.78% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=false/key-types=points-and-ranges-24        147.8n ±  0%   144.8n ±  0%   -2.03% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=true/key-types=points-only-24               144.1n ±  0%   141.7n ±  0%   -1.63% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=true/key-types=points-and-ranges-24         148.2n ±  0%   145.3n ±  0%   -1.99% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=false/key-types=points-only-24              217.3n ±  0%   215.9n ±  0%   -0.64% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=false/key-types=points-and-ranges-24        220.8n ±  0%   219.6n ±  0%   -0.57% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=true/key-types=points-only-24               216.0n ±  0%   214.5n ±  0%   -0.72% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=true/key-types=points-and-ranges-24         220.4n ±  0%   218.1n ±  0%   -1.00% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=7/value-blocks=false/key-types=points-only-24              270.0n ±  0%   270.2n ±  0%        ~ (p=0.146 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=7/value-blocks=false/key-types=points-and-ranges-24        274.1n ±  0%   275.1n ±  1%   +0.36% (p=0.013 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=7/value-blocks=true/key-types=points-only-24               271.2n ±  0%   268.6n ±  0%   -0.96% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=7/value-blocks=true/key-types=points-and-ranges-24         275.6n ±  0%   272.3n ±  0%   -1.20% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=10/value-blocks=false/key-types=points-only-24             301.5n ±  0%   301.6n ±  0%        ~ (p=0.322 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=10/value-blocks=false/key-types=points-and-ranges-24       306.2n ±  0%   305.2n ±  0%   -0.36% (p=0.001 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=10/value-blocks=true/key-types=points-only-24              303.1n ±  0%   298.6n ±  0%   -1.48% (p=0.000 n=10)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=10/value-blocks=true/key-types=points-and-ranges-24        307.3n ±  0%   303.4n ±  0%   -1.29% (p=0.000 n=10)
…